moving graph_line_graph to a separate package
